requirements.txt
flask==2.0.3
tensorflow>=2.6.0
pillow>=8.0.0
numpy>=1.19.0
pandas>=1.3.0
matplotlib>=3.3.0
seaborn>=0.11.0
scikit-learn>=0.24.0
werkzeug==2.0.3
gunicorn>=20.0.0
python-dotenv>=0.19.0
requests>=2.25.0
